add deployment checks (#404)
* added deployment checks job, added pod lib lint

* added deployment checks to circleCI config, added carthage archive step to deployment checls

* added check to ensure that a release in github with the current version number doesn't already exist.

* added extra check to ensure that the git tag doesn't exist before deploying

* bumped xcode version for deployment checks job

* updated xcode version for all jobs

* replaced relative url with link so it works locally as well as in CI

* added explicit casts from NSUInteger to unsigned long in a couple of places

* re-added other deployment checks

* added filter to only run deployment checks on release tags and PRs

+++ Purchases/SubscriberAttributes/RCPurchases+SubscriberAttributes.m
@@ -24,7 +24,7 @@ NS_ASSUME_NONNULL_BEGIN
     NSString *appUserID = self.appUserID;
     RCSubscriberAttributeDict unsyncedAttributes = [self.subscriberAttributesManager
                                                     unsyncedAttributesByKeyForAppUserID:appUserID];
-    RCLog(@"found %lu unsynced attributes for appUserID: %@", unsyncedAttributes.count, appUserID);
+    RCLog(@"found %lu unsynced attributes for appUserID: %@", (unsigned long)unsyncedAttributes.count, appUserID);
     if (unsyncedAttributes.count > 0) {
         RCLog(@"unsynced attributes: %@", unsyncedAttributes);
     }
